neuron blocking detector

+++ src/neurons.h
@@ -16,6 +16,6 @@ float neuron_sense( environment* env, brain* br, int8_t gene_index );
 float neuron_cognition( int8_t gene_index, float input );
 
 /* Active the specified motor neuron */
-void neuron_motor( environment* env, brain* br, int8_t gene_index, int input );
+void neuron_motor( environment* env, brain* br, int8_t gene_index, float input );
 
 #endif
